more problems solved

+++ src/main/java/PossibleMaximum.java
@@ -4,29 +4,29 @@ public class PossibleMaximum {
 
 	public static void main(String[] args) {
 		Scanner s = new Scanner(System.in);
-		int t = s.nextInt();
-		for (int i = 0; i < t; i++) {
-			int n = s.nextInt();
-			int k = s.nextInt();
+		long t = s.nextLong();
+		for (long i = 0; i < t; i++) {
+			long n = s.nextLong();
+			long k = s.nextLong();
 			System.out.println(possibleMaximun(n, k));
 		}
 		s.close();
 	}
 
-	static int possibleMaximun(int n, int k) {
+	static long possibleMaximun(long n, long k) {
 		return findMaximun(0, n, k, 
-				(int j, int s, int e, int x) -> possibleMaximumForEachNumber(j, e, x) );
+				(long j, long s, long e, long x) -> possibleMaximumForEachNumber(j, e, x) );
 	}
 
-	static int possibleMaximumForEachNumber(int start, int end, int k) {
+	static long possibleMaximumForEachNumber(long start, long end, long k) {
 		return findMaximun(start, end, k, 
-				(int j, int s, int e, int x) -> s&j );
+				(long j, long s, long e, long x) -> s&j );
 	}
 
-	static int findMaximun(int start, int end, int k, GetValue getValue) {
-		int max = 0;
-		for (int i = start+1; i <= end; i++) {
-			int possibleValue = getValue.done(i, start, end, k);
+	static long findMaximun(long start, long end, long k, GetValue getValue) {
+		long max = 0;
+		for (long i = start+1; i <= end; i++) {
+			long possibleValue = getValue.done(i, start, end, k);
 			if(max < possibleValue) {
 				max = possibleValue;
 			}
@@ -38,7 +38,7 @@ public class PossibleMaximum {
 	}
 
 	static interface GetValue {
-		int done(int i, int start, int end, int k);
+		long done(long i, long start, long end, long k);
 	}
